- (random.h, random.c, main.c, autoconf/configure.in, config.h.in, efuns.c,     random/*)   Exchanged the Mersenne Twister (pseudo random number generator) by the     SIMD-oriented Fast Mersenne Twister (SFMT) to support random numbers up     to 2^63-1 on 64 bit platforms:     http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/SFMT/     Additionally, the period length can be configured now. The SFMT     faciliates a better equidistribution and a faster recovery from a     0-excess state as well. (Bug #527)   Added the possibility to seed the PRNG from arbitrary files or devices by     using the commandline option --randomdevice. The default device is     /dev/urandom. If that is not readable, the driver falls back to seed the     PRNG with the system time.
git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2470 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/efuns.c
@@ -8738,7 +8738,7 @@ f_random (svalue_t *sp)
     if (sp->u.number <= 0)
         sp->u.number = 0;
     else
-        sp->u.number = (p_int)random_number((uint32)sp->u.number);
+        sp->u.number = (p_int)random_number(sp->u.number);
 
     return sp;
 } /* f_random() */
